tsp-backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
black==21.12b0
click==8.0.3
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
